refactor: split server.py, extract shared helpers, log silent handlers
Second and final pass of the complexity-reduction work. Stacked on the
graph-layer cleanup in the parent branch.

server.py: 9,619 -> 8,243 (-1,376, -14.3%)

Three modules extracted following the existing graph_tools.register()
pattern, plus one shared helper module:
  inv_store.py            manifest/JSONL storage helpers (_now, _inv_dir,
                          _load_manifest, _save_manifest, _append_jsonl,
                          _read_jsonl, _summarise_finding, _make_ref) --
                          zero definitions remain in server.py
  investigation_tools.py  investigation_* tools, attached via register()
  llm_tools.py            local-model offload tools
  model_json.py           shared model-output JSON parsing

Tool registration is now 41 decorated in server.py plus 30 attached by
three register() calls. Total is still 71, names unique, and every one
re-exported on the server module -- verified, not assumed.

Silent exception handlers across mcp/ (non-test): 90 -> 44. Each gained a
diagnostic; fail-open behaviour is untouched -- every pass, continue and
fallback value is exactly as before.

Also: removed 58 lines of uncallable code in memcheck, collapsed two
43-line duplicate blocks behind a differential test, and added 16 tests to
memcheck/code_rules, which had none.

Verified: 487 passed / 0 skipped (up from 467; the +20 are the two new test
files, confirmed in isolation, so no existing test was lost or silently
skipped). Ruff gate clean, including when extended to every new module.
